Implemented the Forgot Password Functionality
On the log in screen you can now press forget password. It will ask for the email of the account. If it finds the email is in our database, it will send an email with a link to a reset. I will get to that a different time. PHPMailer-master is what allows emails to be sent from our akari.test496@gmail.com