Fixed notification bugs, started flag system
Notifications are no longer buggy. When you click on a notificaiton, only that one will become read and will then show the correct amount of remaining unread notifications

Started flagging process. Clicking flag on properties page will increment the flag count for that property using AJAX. If it is over 5, it will send an email to the email we set up (akari.test296@gmail.com). So far this is working fine