Styles chooseRenter page, changed loginsignup to php
I reformated the way that the user selects the property they want to rent out. Just looks nicer

If a user was logged in, they could type in the address par and get to the login page again, even though they were logged in. I needed to restrict this with PHP so I had to change the file name and the name everywhere it was linked to throughout the other files.

Also added mobile PHP files for the mobile application. Updated listing details, register, and register property to accomadate for database changes. Also added upload image php file to upload an image from the mobile phone taken with the camera