Redesigned profile page, send emails to renter and landlord when rental is created
Profile page redesign:
Added a portion so that a user that is renting a property can view it. Before, there was no visual confirmation. There is now two sections. Properties that are owned and properties that you are renting.

When a landlord completes tenency, it sends a personalized email to each party with details about the rental. It will then send a notification to the renter's profile from the landlord thanking them for renting the property

Corrected spelling of confirm php and fixed that spelling everywhere else