Fixed image issue for properties page
Images are responsive and dont leave gaps in the list of properties (from my testing on 27" iMac, may be different on other screens).

This is not the case for the profile page. I could make them work the same way but the reason it doesnt work for profile page is because we are trying to make the images bigger. If they were the same size as properties page, it would work